refactor: `wow.parallelism` for `MessageDispatcher`
