+++ gradle.properties
@@ -15,7 +15,7 @@ ksp.incremental=true
 ksp.incremental.log=true
 
 group=me.ahoo.wow
-version=1.8.23
+version=1.8.24
 description=A Modern Reactive CQRS Architecture Microservice development framework based on DDD and EventSourcing.
 website=https://github.com/Ahoo-Wang/Wow
 issues=https://github.com/Ahoo-Wang/Wow/issues
