+++ wow-core/src/main/kotlin/me/ahoo/wow/event/AbstractEventDispatcher.kt
@@ -16,7 +16,6 @@ package me.ahoo.wow.event
 import me.ahoo.wow.api.modeling.NamedAggregate
 import me.ahoo.wow.messaging.MessageDispatcher
 import me.ahoo.wow.messaging.dispatcher.AbstractDispatcher
-import me.ahoo.wow.messaging.dispatcher.MessageParallelism
 import me.ahoo.wow.messaging.writeReceiverGroup
 import me.ahoo.wow.metrics.Metrics.writeMetricsSubscriber
 import reactor.core.publisher.Flux
@@ -24,7 +23,7 @@ import reactor.core.publisher.Mono
 import reactor.core.scheduler.Scheduler
 
 abstract class AbstractEventDispatcher<R : Mono<*>> : AbstractDispatcher<EventStreamExchange>() {
-    abstract val parallelism: MessageParallelism
+    abstract val parallelism: Int
     abstract val domainEventBus: DomainEventBus
     abstract val functionRegistrar: AbstractEventFunctionRegistrar<R>
     abstract val eventHandler: EventHandler
