+++ wow-core/src/main/kotlin/me/ahoo/wow/saga/stateless/StatelessSagaDispatcher.kt
@@ -26,7 +26,7 @@ class StatelessSagaDispatcher(
      * named like `applicationName.StatelessSagaDispatcher`
      */
     override val name: String,
-    override val parallelism: MessageParallelism = MessageParallelism.DEFAULT,
+    override val parallelism: Int = MessageParallelism.DEFAULT_PARALLELISM,
     override val domainEventBus: DomainEventBus,
     override val functionRegistrar: StatelessSagaFunctionRegistrar,
     override val eventHandler: StatelessSagaHandler,
